api/cart/requirements.txt
flask==2.1.1
flask-restful==0.3.9
pymongo==4.1.0
requests
PyJWT